+++ R/ebv_resample.R
@@ -80,6 +80,7 @@ ebv_resample <- function(filepath_src, datacubepath_src = NULL, entity_src=NULL,
                          scenario = NULL, metric = NULL, return_raster=FALSE,
                          overwrite = FALSE, ignore_RAM=FALSE, verbose=TRUE){
   ####initial tests start ----
+
   # ensure file and all datahandles are closed on exit
   withr::defer(
     if(exists('hdf')){
@@ -305,7 +306,12 @@ ebv_resample <- function(filepath_src, datacubepath_src = NULL, entity_src=NULL,
   #align to origin of the destination file
   data_proj <- tryCatch(
     {
-      data_proj <- terra::project(data_ts, y = dummy, align=TRUE, method=method, gdal=TRUE)
+      data_proj <- terra::project(data_ts, y = dummy,
+                                  align = TRUE,
+                                  method = method,
+                                  gdal = TRUE,
+                                  use_gdal = TRUE,
+                                  threads = TRUE)
     },
     error=function(e){
       # if (!stringr::str_detect(e, 'cannot create dataset from source')){
@@ -315,7 +321,12 @@ ebv_resample <- function(filepath_src, datacubepath_src = NULL, entity_src=NULL,
         message('Slower algorithm needs to be used. Please be patient.')
       }
 
-      data_proj <- terra::project(data_ts, y = dummy, align=TRUE, method=method, gdal=FALSE)
+      data_proj <- terra::project(data_ts, y = dummy,
+                                  align = TRUE,
+                                  method = method,
+                                  gdal = TRUE,
+                                  by_util = TRUE,
+                                  threads = TRUE)
     }
   )