+++ R/ebvcube-package.R
@@ -2,13 +2,14 @@
 #'
 #'
 #' This package can be used to easily access the data of the EBV netCDFs which
-#' can be downloaded from the \href{https://portal.geobon.org}{Geobon Portal}. It also
-#' provides some basic visualization of the data.  Advanced users can build
-#' their own netCDFs with the EBV standard using this package.
+#' can be downloaded from the \href{https://portal.geobon.org}{EBV Data Portal}.
+#' It also provides some basic visualization of the data.  Advanced users can
+#' build their own netCDFs in the EBVCube format using this package.
 #'
-#' This package contains three main usecases: accessing the data, visualising
-#' it and creating your own data in the EBV netCDF
-#' standard.
+#' This package contains three main usecases: accessing the data, visualising it
+#' and creating your own data in the EBVCube netCDF format. See the
+#' \href{https://github.com/EBVcube/EBVCube-format/}{EBVCube format repository}
+#' for a detailed description of the
 #'
 #'
 #' @name ebvcube